Edited MealPlanGeneratorView to have a button that takes the user to their saved meal plans, alongside an error message if the user does not have any saved meal plans.
Also changed logic for the Navbar so that when you tab away to a different view and tab back to the meal plan generator, all the fields would be reset. This required re-ordering the build order when the AppBuilder is called upon in Main, alongside allowing the Navbar Presenter to take in a new MealPlanGeneratorViewModel argument.